Add support for the XMPP Ping protocol.
Author: ralphm
Fixes #55.

This adds a whole bunch of examples around the ping protocol in different
settings for both the pinger and pingee: as a external server-side component,
as a standalone server with server-to-server connectivity and as a client.

Also updates the disco support to allow for non-deferred responses from
subprotocol handlers

--HG--
extra : convert_revision : svn%3Ab33ecbfc-034c-dc11-8662-000475d9059e/trunk%40183