Make sure item and subscription elements get the correct namespace.
This complements the changes in [f4aa0d507bc8] and [3fe44cf07366] for item and
subscription elements that are used in different namespaces. These changesets
set the namespace to `None`, which works nicely when the elements are
serialized. However, in situations where the DOM representation is passed
around (in-process communication) these elements are no longer matched by
namespace. This change makes sure the elements get the proper namespace just
before they are sent out.