Add iq request (set/get) tracking for StreamManager.
This adds a new `request` method to StreamManager to serialize an iq get or iq
set request object (using toElement), send it over the stream (or queue it
until the next time a stream is initialized) and track the response stanza by
returning a Deferred.

Requests may have a timeout, so that the deferred errbacks after the timeout
(counting from the moment of calling `request`) has expired.

This implementation is similar to the response tracking done in
`twisted.words.protocols.jabber.xmlstream.IQ`, but slighly different:

  * Requests can be queued until the connection is initialized, even before
    a XmlStream object is available. I.e. it is tied to the StreamManager, not
    the stream.
  * Response timeout starts counting from the moment of calling `request`, not
    from the moment request is sent over the wire.