Redo admin requests using generic.Request.
This changes all role and affiliation management functionality to use a
subclass of generic.Request.

The methods for retrieving affiliation or role lists now return lists of
`AdminItem`s, instead of modifying a `Room` instance and returning that. This
is because affiliation changes can also be done when not in the room.

--HG--
branch : wokkel-muc-client-support-24