Split protocol and room/user administration into two classes.
This introduces `MUCClientProtocol`, which only contains the protocol
implementation. The only state keeping there is remembering the Occupant JID
that was used to join a room, so you can use the Room JID everywhere.

The room and user owner administration that was intertwined with the protocol
is still in `MUCClient` that is a subclass of `MUCClientProtocol`. As this
record keeping barely had proper tests, the test coverage for this class is
sub-optimal for now.

This also fixes various test cases for the pure protocol support.

--HG--
branch : wokkel-muc-client-support-24