Expose locked state of a room on join response with room created status.
When the presence in response to a join includes the room created status code
(201), the client should configure the room before it can be used. To enable
this, the locked state of the room is now exposed through the `Room.locked`
attribute.

Also, update example MUC client to deal with newly created, locked rooms.