Make sure 'twisted' doesn't appear in the egg-info top_level.txt file.
Wokkel provides a plugin for twistd. As with all twisted plugins, these
need to be placed in a directory twisted/plugins so that the plugin mechanism
can find them. This is not a real package, and when building eggs the
directory must appear in the egg. However, 'twisted' should not occur in the
top_level.txt file as this causes warnings when importing twisted, and
triggers a bug in pip that removes all of twisted when a package with a
twisted plugin is removed.

This change overrides the way the top_level.txt file is written, by explicitly
excluding 'twisted'.

Author: ralphm.
Fixes: #76.